nav/pyproject.toml
[project]
name = "nav"
version = "0.1.0"
description = "Helper for generating a 3d path from coordinates using google maps and publishing to foxglove over a websocket"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "foxglove-sdk>=0.16.6",
    "googlemaps>=4.10.0",
    "polyline>=2.0.4",
    "pyproj>=3.7.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scipy>=1.17.0",
]
